requirements.txt
Flask
Flask-RESTful
python-gammu



